Refactor agent execution system for better modularity and maintainability
Major code organization improvements with no breaking changes:

**New modules created:**
- agent/constants.py: Extracted AGENT_PARAMS and MESSAGE_PARAM constants
- agent/context_types.py: Moved AgencyContext dataclass for type organization
- agent/execution_guardrails.py: Isolated guardrail logic with enhanced metadata
- agent/execution_streaming.py: Separated streaming execution with improved handoff detection

**Core improvements:**
- Reduced agent/core.py by ~50 lines, removed constants and type definitions
- Shrunk execution_helpers.py from 670 to 327 lines by extracting streaming logic
- Refactored console_event_adapter.py with private handler methods for better organization
- Enhanced guardrail feedback with proper message_origin classification
- Improved handoff detection supporting both legacy and new event formats

**Benefits:**
- Better separation of concerns across agent execution system
- Reduced file sizes for improved maintainability
- Enhanced error handling and event processing
- Preserved all existing APIs and functionality
- Cleaner import structure and dependencies
